QBenchmarkPerfEventsMeasurer: move the perf_event_attr data to the stack
There's no need for it to be a global variable. It needed to be global
because we configured the event globally before start(), but in commit
4731baf6d3a18857e86cc16de000bc42e84bf6de (6.5.0) we instead introduced
the Q_GLOBAL_STATIC with a QList containing the globally-configured
events.

This is using designated initializers despite their being a C++20
feature, because GCC and Clang have supported them as an extension to C+
+ for a long time. This file has needed that extension since that commit
anyway.

Pick-to: 6.8
Change-Id: I1c9080d23df9f7ba3cbafffd7eae7bd57cd69b67
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>