Fix duplicate datatags and other issues in tst_QSizePolicy
tst_QSizePolicy is creating data tags programmatically, and the
implementation creates duplicates. This commit fixes the problem.

It used to print tens of thousands of warnings about duplicate datatags:

QWARN  : tst_QSizePolicy::getSetCheck() Duplicate data tag "Fixed-Fixed-0-0-DefaultType-false-false" - please rename

It turns out that the warnings were not only caused by collisions of
datatags (the textual representations of the test data rows), but also
by real test issues.

Problems fixed or found:

The datatag string now includes a couple of new factors used in
generating the data and were not represented before. Namely
"ed" (expandingDirections), and "spId" which represents the counter `i`
from the innermost loop which changes what each iteration of the test
checks for.

That same innermost counter was previously iterating up to one value
less than it should (4 instead of 5 (inclusive)), which resulted to test
data rows containing the same values.

The very first iteration of the innermost loop is now skipped so that
"oldsp" gets a real value, which avoids 5 data rows identical to the
ones defined on the next iteration.

Mark with a FIXME comment that the part of the test filling the
expandingDirections data does nothing, because none of the "policies"
tested has ExpandFlag set. The QSizePolicy-ies that have that, are
Expanding and MinimumExpanding, which are not tested at all. In other
words, all generated data have no expanding directions, thus the
datatags contain the "ed:__" string.

Finally, the datatags format string is now more descriptive, of the
form:

tst_QSizePolicy::getSetCheck(H:Fixed-V:Fixed-0-0-DefaultType-hfw:0-wfh:0-ed:__-spId:0)

Task-number: QTBUG-107185
Change-Id: I48d5839967f743643c68aeb0f3950503cd5a6e0e
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>