a11y: Report QMessageBox accessible ID
Commit b8b7c58402740204da72e1b1f4ea7321b7bfa540 introduced
QWidget::setAccessibleIdentifier for setting the accessible
ID of a widget.

When calling that on a QMessageBox, the ID set this way would
however not get reported on the platform level, e.g.
AT-SPI2 on Linux.

This is because QAccessibleMessageBox overrides
the default implementation of QAccessibleWidget::text
and so far didn't explicitly handle QAccessible::Identifier,
but returned an empty string for the default case instead.

Call the base class implementation for the default
case to make this work and also handle other currently
unhandled cases (Accelerator, DebugDescription), and
potential QAccessible::Text enum values that might
get added in the future.

While at it, unify to call the base class implementation
for QAccessible::Description as well instead of manually
calling QWidget::accessibleDescription. That is what the
base class implementation does as well, and it includes
an additional fallback to use the tooltip if no accessible
description is explicitly set.

Change-Id: I0f3c53624b1c71ca666f836ba353f1bb1da9a8bc
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>