Rework the argument escaping and wrapping in _qt_internal_create_command_script
Use the quote-based argument wrapping when generating the command line
scripts for the tests. Apply the wrapping and escaping of the arguments
right before writing the script with the following advanced rules:
- if argument is already wrapped using quotes, skip wrapping
- if argument is a CMake variable, skip wrapping
- wrap the argument with qoutes otherwise

The above should cover the most common usecases in centralized place
when we generate the test wrapper scripts.

Remove other ways we wrap arguments in other places, like square bracket
wrapping in qt_internal_create_test_script.

Change-Id: If287dd75d6fb36260b5cf8a687215bda9c9fc1c0
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>