Extract toQtModifiers to place where it belongs
`QEvdevKeyboardHandler` contains static method `toQtModifiers` which is
not operating on any of its internal fields or types. This indicates
that this method doesn't belong to this class, and in fact it should be
moved to `QEvdevKeyboardMap` where types used in it are declared.

Move this static method to `QEvdevKeyboardMap` namespace. This is a
preparatory change to extract `QEvdevKeyboardMap` to separate header.

Task-number: QTBUG-115777
Change-Id: Ibe88b40a56fb96772d903498611777b3ec19dded
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>