QTextCursor: Consider QTextBlock's layout text direction
QTextCursor::movePosition uses QTextBlock::textDirection in order to
decide how to convert "left" and "right" move operations into an
actual direction (back/forward). However the heuristics in that method
are not necessarily those used by the QAbstractTextDocumentLayout
subclass that's rendering the document.

This makes cursor movements inconsistent with the with actual text
rendering and cursor shape as done by QTextEngine. To fix this,
prefer to use the text direction set on the QTextBlock's QTextOption
(which is what QTextEngine uses), using the previous method only as
a fallback.

Fixes: QTBUG-130317
Pick-to: 6.11
Change-Id: I80f27c9c6d34601f250ef008cdd720a09ed75097
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>