QUtf8: simplify the AVX2 code in simdDecodeAscii
Instead of having this extra code to deal with BitSpacing because we
were getting the bit mask from the UTF-16 content, we can get it from
the UTF-8 side.

This means the compilers must emit one extra instruction, but the number
of uops to be executed is roughly the same. See the LLVM-MCA analysis:
https://analysis.godbolt.org/z/TP48jM9Tq

Change-Id: I73641d6b9443d8216eadfffd71515d1f33b3d833
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>