Fix transition times for the new tzdb backend to QTimeZone
The nextTransition() and previousTransition() implementations used the
time they were passed as the atMSecsSinceEpoch of the returned object,
leading to infinite looping in a caller looking for a different
transition.

In nextTransition(), the conversion to fromSysInfo() got *info where
it should have got the info for tran, so fix that in the process.

In the process, fix a conflict resolution glitch that left the old
form of the MS-without-ICU check in tst_QDateTime.

Change-Id: Ib21e58fae6c3a8a0dbd10339b19492a1a2d78b08
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Magdalena Stojek <magdalena.stojek@qt.io>