QUtf8: modernize simd{Encode,Decode}Ascii()
This rewrites the looping so we only enter that block if we have 16 or
more characters to process. Moreover, we only loop if there are more
than 16 characters in the first place, otherwise we execute a non-
looping block to handle exactly 16 characters. Unfortunately, the
presence of that block after the loop causes both Clang and GCC to emit
slightly worse code for the loop (LLVM-MCA simulations show no change
for long loops on ADL-P but cost 1 extra cycle on SKL).

We also process in vector mode anything with 4 characters or more,
leaving the scalar code to deal with only up to 3. Those two are
implemented by a pair of overlapping loads and stores. Like the loop
above, we only perform an overlapping store if the previous block had no
non-ASCII character.

For simdEncodeAscii(), We use a slightly ugly calculation of size in
bytes, to avoid the 1 byte gap caused by char16_t being 2 bytes. Clang
generates the exact same code for AVX and AXV2 for this code. The
LLVM-MCA analysis[1] shows it runs at 2.0 cycles/iteration for 16
characters per iteration (8 characters/cycle).

In this simdDecodeAscii() implementation, according to the LLVM-MCA
analysis[2] the non-AVX code (compiled with AVX) runs at 2.3
cycles/iteration, processing 16 characters (~7.0 characters/cycle),
while the AVX2 implementation only needs 1.5 cycles/iteration (10.5
characters/cycle).

Drive-by C++ modernization too.

[1] https://analysis.godbolt.org/z/h5K4xzzW4
[2] https://analysis.godbolt.org/z/3Pfnso787

Change-Id: I19ccdb6485843dd9f206fffd4bd468264c66ace4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>