CMake: Improve doc file copying dependency handling in qtbase
If doc/global/macros.qdocconf were modified, rerunning the doc
generation targets in qtbase or a top-level build would use the old
non-modified contents, because the file is only copied into the build
dir at configure time, and there is no dependency to rerun configure.

Add all the qtbase doc config files to CMAKE_CONFIGURE_DEPENDS, so
that configure is reran and the files are re-copied if the files are
touched.

Add the QT_NO_SET_QTBASE_DOC_FILE_DEPS_COPYING variable as an opt out,
in case this breaks someone's workflow in an unexpected way where they
have to reconfigure too often.

This is an incremental improvement over no dependencies at all.

Investigating whether setting up fine-grained build-time dependencies
is possible, is left for the future.

Pick-to: 6.5 6.8
Change-Id: I8f0702aa789b5116ed63a3a78c5d3b19b2177131
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>