DecompressHelper: retain the error from internal counter
We never copied the error string so after the call to 'clear' it went
away completely. Additionally, stop clearing the error string in clear()
since we reset it when the encoding is set, in case the object is
reused.

Pick-to: 6.5
Fixes: QTBUG-129697
Change-Id: Ia64e1d13a99b62760f61cac6b67ae3cff5e2c9da
Reviewed-by: Mate Barany <mate.barany@qt.io>
(cherry picked from commit f84c83be52462536ce5ae5ba68205e7061369fe3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>