Add systemTimeZoneId() to the C++20 tzdb backend of QTimeZone
It's not strictly necessary, but it should be a little more
efficient. Without it, QTZ::systemTimeZoneId() falls back to using the
id() of a default-constructed backend, QChronoTZP(), which will get
the same answer returned here, but we can avoid the need for that
fall-back by doing the right lookup to begin with.

Change-Id: I6fe5beba048b5e9817b6bfed6eba785b13c2c206
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>