Add a helper CMake script to bring Windows App SDK to windows apps
This CMake file will be included in the user's project.
This script adds a function called "qt_add_win_app_sdk".
This function needs to be called in the user's project's CMake file.
This is the first step to bring Windows App SDK functionality to Qt
apps and it might be enough. For now there is not enough reason to
integrate Windows App SDK into Qt. Title bar customization can be done
without needing the SDK, and WebView2 can also be used independently.

Task-number: QTBUG-124800
Change-Id: Icd4ca2c51783942cbb8bfe1c274162628b92a115
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>