QTestLib: increment timestamp by specified delay before each event
Some test functions generate series of mouse events, and the docs have
always said that the specified delay applies to each event, not just
to the first one. This brings reality into agreement with the docs.

We're still keeping 0f94430a0fba5509d2d756832cb44a266d1d2e86 : the
delay cannot be 0, because if we never increment the timestamp at all,
that would never be a realistic simulation of any real-world scenario.
Checking elapsed time is increasingly useful in Qt, and unrealistic
timestamps can cause strange behavior.

[ChangeLog][QTestLib] QTest::mouseClick() and mouseDClick() now apply
the given delay to the timestamps of each event in the sequence, as it
has been documented all along. As before, the delay is never allowed
to be less than 1 ms.

Fixes: QTBUG-129794
Change-Id: I03de902b57582de239fc9491b95fb406b879f7e3
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>