Add more debug messages to the lldb QtCore debug_script.py
So that it's easier to troubleshoot why a certain bridge might not be
loaded.

To enable verbose debug messages, set the
QT_LLDB_SUMMARY_PROVIDER_DEBUG environment variable to 1 when
launching Qt Creator or lldb.

The informational message to set the env var is also shown if no
bridge was loaded successfully.

Pick-to: 6.8
Task-number: QTCREATORBUG-31769
Change-Id: I53e7b26950fdd635e51d0458cb5fb90600b158a2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>